fix deploy.sh

+++ deploy.sh
@@ -6,20 +6,24 @@ mkdir -p arhives
 mv myproject arhives/myproject.${BUILD_NUMBER}
 
 echo 'Stopping uwsgi..'
-echo 'processes', $(ps -ef | grep [w]sgi | awk '{print $2}')
-kill $(ps -ef | grep [w]sgi | awk '{print $2}')
+uwsgi = $(ps -ef | grep [w]sgi | awk '{print $2}')
+
+echo 'processes' $uwsgi
+if [ -n $uwsgi ];then
+    kill $(ps -ef | grep [w]sgi | awk '{print $2}')
+    sleep 1
+fi
 
 echo 'Unzipping project..'
 unzip -o backend.zip -d myproject
 mv myproject/goldpoisk/settings.prod.py myproject/goldpoisk/settings.py
+sleep 1
 
 echo 'Migrations..'
 echo 'Migration product..'
 cp -r arhives/myproject.${BUILD_NUMBER}/goldpoisk/product/migrations myproject/goldpoisk/product/
-cd myproject
-python manage.py makemigrations product -v 2
-python manage.py migrate product -v 2
-cd ..
+python myproject/manage.py makemigrations product -v 2
+python myproject/manage.py migrate product -v 2
 
 rm backend.zip
 echo 'Restart uwsgi && nginx..'
